# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/post/{id}:
    get:
      tags:
      - posts-controller-rest
      operationId: getIndividualPost
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Post_PostDetail'
    put:
      tags:
      - posts-controller-rest
      operationId: editPost
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Post'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Post'
    delete:
      tags:
      - posts-controller-rest
      operationId: deletePost
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Post'
...
  /api/chat/{id}:
    get:
      tags:
      - chat-controller-rest
      operationId: getIndividualChat
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Chat_ChatDetail'
    put:
      tags:
      - chat-controller-rest
      operationId: addMessage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: message
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Chat_ChatDetail'
    delete:
      tags:
      - chat-controller-rest
      operationId: deleteChat
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Chat_ChatDetail'
...
  /api/chat/:
    post:
      tags:
      - chat-controller-rest
      operationId: createChat
      parameters:
      - name: gameID
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Chat_ChatDetail'
...
components:
  schemas:
...
    Chat:
      type: object
      properties:
        listMessages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        ID:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
    Game:
      type: object
      properties:
        id:
          type: integer
          format: int64
        gameTitle:
          type: string
        genre:
          type: string
          enum:
          - Horror
          - Shooter
          - Action
          - Platformer
          - Sports
          - Puzzles
          - Narrative
          - RPG
        mapScores:
          type: object
          additionalProperties:
            type: integer
            format: int32
        averageScore:
          type: number
          format: float
        description:
          type: string
        chat:
          $ref: '#/components/schemas/Chat'
        thePosts:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        imagePath:
          type: string
    Message:
      type: object
      properties:
        nameUser:
          type: string
        msgString:
          type: string
        id:
          type: integer
          format: int64
          writeOnly: true
        messageWriter:
          type: boolean
    Post:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        creationDate:
          type: string
        updateDate:
          type: string
        fromGame:
          $ref: '#/components/schemas/Game'
        postText:
          type: string
        author:
          type: string
        theType:
          type: string
          enum:
          - News
          - Updates
          - Guides
        imagePath:
          type: string
...
    Chat_ChatDetail:
      type: object
      properties:
        listMessages:
          type: array
          items:
            $ref: '#/components/schemas/Message_ChatDetail'
        ID:
          type: integer
          format: int64
        id:
          type: integer
          format: int64
    Message_ChatDetail:
      type: object
      properties:
        nameUser:
          type: string
        msgString:
          type: string
        id:
          type: integer
          format: int64
          writeOnly: true
        messageWriter:
          type: boolean
...
    Chat_PostDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
    Game_PostDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        gameTitle:
          type: string
        genre:
          type: string
          enum:
          - Horror
          - Shooter
          - Action
          - Platformer
          - Sports
          - Puzzles
          - Narrative
          - RPG
        mapScores:
          type: object
          additionalProperties:
            type: integer
            format: int32
        averageScore:
          type: number
          format: float
        description:
          type: string
        chat:
          $ref: '#/components/schemas/Chat_PostDetail'
        thePosts:
          type: array
          items:
            $ref: '#/components/schemas/Post_PostDetail'
        imagePath:
          type: string
    Post_PostDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        creationDate:
          type: string
        updateDate:
          type: string
        fromGame:
          $ref: '#/components/schemas/Game_PostDetail'
        postText:
          type: string
        author:
          type: string
        theType:
          type: string
          enum:
          - News
          - Updates
          - Guides
        imagePath:
          type: string